requirements.txt
django
gunicorn
psycopg2
south
django-registration
djangorestframework
https://github.com/bartTC/django-attachments/zipball/master
